paxctl: new package.
paxctl is a tool that allows PaX flags to be modified on a per-binary
basis.

PaX is a patch for Linux which provides memory pages protection, Address
Space Layout Randomization and more. It is used to harden the kernel. It
is provided by Grsecurity.

Your system needs to be running a properly patched and configured kernel
for this program to have any effect.

See http://grsecurity.net for more details.

Signed-off-by: Eric Le Bihan <eric.le.bihan.dev@free.fr>