libwayland-egl: new virtual package
Currently, when wayland and EGL support are selected, mesa3d builds the
libwayland-egl library, which allows Wayland to use the libEGL
implementation provided by mesa3d.

But some embedded hardware devices like Raspberry Pi provides their own
implementation of libEGL. So it is likely that an implementation of
libwayland-egl for RPi will appear sooner or later.

So a new virtual package called libwayland-egl is added, to handle these
future cases.