package/dub-bin: new package
Dub is the package and build management system for the D programming language.

As it is written in D, it requires a D compiler on the host machine. Since this
is not the default on GNU/Linux systems, it is not possible to build the host
variant from sources.

This new package provides the binary host variant.